Begin using ZSH and modular install

+++ install_ubuntu.sh
@@ -1,19 +1,33 @@
 #!/usr/bin/env bash
 
-git clone git://github.com/jimeh/git-aware-prompt.git
+sudo apt install fonts-powerline zsh
+sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+git clone https://github.com/zsh-users/zsh-syntax-highlighting.git
 
-./set_perms.sh
+rm -rf ~/.zshrc
+ln -s ~/repos/dot-files/.zshrc ~/.zshrc
 
-ln -s ~/repos/dot-files/.bashrc ~/.bash_aliases
-
-ln -s ~/repos/dot-files/.gitconfig ~/.gitconfig
 ln -s ~/repos/dot-files/.tmux.conf.linux ~/.tmux.conf
 
-mkdir ~/.gnupg/
-chmod 700 ~/.gnupg/
-ln -s ~/repos/dot-files/gpg.conf ~/.gnupg/gpg.conf
-ln -s ~/repos/dot-files/gpg-agent.conf ~/.gnupg/gpg-agent.conf
+for i in "$@"
+do
+    case $i in
+	--gpg)
+            bash install_gpg.linux.sh
+	    touch .use_gpgrc
+	    shift
+	    ;;
+	*)
+	    ;;
+    esac
+done
 
-sudo apt-get install -y gnupg2 gnupg-agent pinentry-curses scdaemon pcscd
-
-gpg2 --recv-key 0x20575b0850e17457
+if [ -e ~/.gitconfig ]; then
+    echo "Skipping .gitconfig, file already exists!"
+else
+    if [ -f ~/repos/dot-files/.use_gpgrc ]; then
+	ln -s ~/repos/dot-files/.gitconfig_gpg ~/.gitconfig
+    else
+	ln -s ~/repos/dot-files/.gitconfig ~/.gitconfig
+    fi
+fi
